requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
wikipedia
pymupdf
langchain-text-splitters
langchain-openai
langchain-google-genai
langchain_google_vertexai
langchain_chroma
langchain-ollama
sentence_transformers
langchain_huggingface
sentence_transformers
faiss-cpu
streamlit
langchain_groq
langserve
fastapi
crewai
crewai_tools